LPS-55110 Fix Tests
Make sure articles are added/removed through the service logic.
@DeleteAfterRunTest is not required for articles when their group already has this annotation.

+++ portal-impl/src/com/liferay/portlet/journal/service/impl/JournalArticleLocalServiceImpl.java
@@ -973,9 +973,13 @@ public class JournalArticleLocalServiceImpl
 
 		// Dynamic data mapping
 
-		ddmTemplateLinkLocalService.deleteTemplateLink(
-			classNameLocalService.getClassNameId(JournalArticle.class),
-			article.getId());
+		if (article.getClassNameId() !=
+				classNameLocalService.getClassNameId(DDMStructure.class)) {
+
+			ddmTemplateLinkLocalService.deleteTemplateLink(
+				classNameLocalService.getClassNameId(JournalArticle.class),
+				article.getId());
+		}
 
 		// Expando
